video-object-detection/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0
supervision>=0.16.0

# Data processing and visualization
matplotlib>=3.5.0
pandas>=1.3.0
seaborn>=0.11.0
Pillow>=8.3.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.62.0
psutil>=5.8.0

# GStreamer Python bindings (will be handled by system install)
# PyGObject>=3.42.0

# Jupyter support (for Colab compatibility)
ipywidgets>=7.6.0
IPython>=7.25.0

# Development tools
pytest>=6.2.0
black>=21.7.0
flake8>=3.9.0